test(core): colocate providerOptions conversion regressions
Move the providerOptions tool-call regression coverage next to the output converter tests and trim the changeset to a concise release note.
